rename data module to core_api module
