+++ src/OSmOSE/core_api/__init__.py
@@ -1,0 +1,3 @@
+from OSmOSE.core_api.audio_file_manager import AudioFileManager
+
+audio_file_manager = AudioFileManager()
