+++ src/OSmOSE/core_api/base_dataset.py
@@ -10,9 +10,9 @@ from typing import TYPE_CHECKING, Generic, TypeVar
 
 from pandas import Timedelta, Timestamp, date_range
 
-from OSmOSE.data.base_data import BaseData
-from OSmOSE.data.base_file import BaseFile
-from OSmOSE.data.event import Event
+from OSmOSE.core_api.base_data import BaseData
+from OSmOSE.core_api.base_file import BaseFile
+from OSmOSE.core_api.event import Event
 
 if TYPE_CHECKING:
     from pathlib import Path
