--- src/OSmOSE/data/__init__.py
@@ -1,3 +1,0 @@
-from OSmOSE.data.audio_file_manager import AudioFileManager
-
-audio_file_manager = AudioFileManager()
