+++ tests/test_spectro.py
@@ -10,12 +10,12 @@ from scipy.signal import ShortTimeFFT
 from scipy.signal.windows import hamming
 
 from OSmOSE.config import TIMESTAMP_FORMAT_EXPORTED_FILES, TIMESTAMP_FORMAT_TEST_FILES
-from OSmOSE.data.audio_data import AudioData
-from OSmOSE.data.audio_dataset import AudioDataset
-from OSmOSE.data.audio_file import AudioFile
-from OSmOSE.data.spectro_data import SpectroData
-from OSmOSE.data.spectro_dataset import SpectroDataset
-from OSmOSE.data.spectro_file import SpectroFile
+from OSmOSE.core_api.audio_data import AudioData
+from OSmOSE.core_api.audio_dataset import AudioDataset
+from OSmOSE.core_api.audio_file import AudioFile
+from OSmOSE.core_api.spectro_data import SpectroData
+from OSmOSE.core_api.spectro_dataset import SpectroDataset
+from OSmOSE.core_api.spectro_file import SpectroFile
 from OSmOSE.utils.audio_utils import generate_sample_audio
 
 if TYPE_CHECKING:
